requirements.txt
appdirs==1.4.0
falcon==1.0.0
hug==2.2.0
hug-middleware-cors==1.0.0
packaging==16.8
pyparsing==2.1.10
python-mimeparse==1.6.0
requests==2.13.0
six==1.10.0
azure.storage==0.36.0
applicationinsights